Change microphysics dtypes to something reasonable

+++ fuse/plugins/micro_physics/find_cluster.py
@@ -23,7 +23,7 @@ class FindCluster(strax.Plugin):
     
     provides = "cluster_index"
     
-    dtype = [('cluster_ids', np.int64),
+    dtype = [('cluster_ids', np.int32),
             ]
     dtype = dtype + strax.time_fields
